Split HookData interface and implementation
